Remove dependency on sudo
Just ensure that rules requiring root privileges are executed as uid=0,
and notify the user if this is not the case. The user should take care
of gaining root privileges, e.g. by envoking the rule with `sudo make
<rule>` or `doas make <rule>`.

This approach provides greater flexibility, and is considered a best
practice.