Follow XDG Base Directory Specification version 0.8
https://specifications.freedesktop.org/basedir-spec/basedir-spec-0.8.html

There is a single base directory relative to which user-specific state
data should be written. This directory is defined by the environment
variable `${XDG_STATE_HOME}`. If `${XDG_STATE_HOME}` is either not set
or empty, a default equal to `"${HOME}/.local/state"` should be used.

`${XDG_STATE_HOME}` contains state data that should persist between
(application) restarts, but that is not important or portable enough to
the user that it should be stored in `${XDG_DATA_HOME}`. It may
contain:

* actions history (logs, history, recently used files, …)
* current state of the application that can be reused on a restart
  (view, layout, open files, undo history, …)