Revert to default UI coloring
By default (`color.ui = auto`), git automatically colors output going to
a terminal, and omits colors when redirecting to a pipe or file. This is
the behavior we want, so we just go with the default.

The documentation mentions possible values `auto` and `always`, so the
previously set value of `true` may have been invalid anyway.