Reduce code duplication
XDG-compliant user-level configuration removal can be simplified, by
creating an implicit rule.

Marking the removal targets phony turns off implicit rule matching, so
we have to make them non-phony. This is fine, because marking a target
as phony accomplishes two things [0]:

1. it prevents conflicts with files matching the target name, and
2. it improves performance, by skipping the implicit rule matching.

We don't have any files matching `remove_*`, so 1. is not a problem, and
we do explicitly want to match implicit rules, so 2. is also what we
want.

[0]: https://www.gnu.org/software/make/manual/make.html#Phony-Targets