Fix zsh syntax highlighting
The zsh syntax highlighting plugin stopped working the way we sourced it
before at some point. Now it has to be sourced at the end of the config
file. It thematically fits nicely into the custom prompt section anyway,
which is already conveniently located at the end of the configuration.